Add files via upload
New Version: Can specify the amplitude range in the waveform.
You cannot have the amplitude units displayed because the label for the lower amplitude limit of the waveform will overlap with the label for the upper frequency limit of the spectrogram. For this reason, it is best to specify the amplitude range from the outset. That way, all sound files that are being drawn will have the same amplitude range.